Add auth policies and conditions to update and delete questions

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -55,7 +59,8 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+        return view('questions.show', compact('question', $question));
     }
 
     /**
@@ -66,6 +71,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if (\Gate::denies('update-question', $question)) {
+        //     abort(403, 'Access denied');    
+        // }
+
+        $this->authorize('update', $question);
+
         return view('questions.edit', compact('question', $question));
     }
 
@@ -78,6 +89,12 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        // if (\Gate::denies('update-question', $question)) {
+        //     abort(403, 'Access denied');    
+        // }
+
+        $this->authorize('update', $question);
+
         $question->update($request->only('title', 'body'));
 
         return redirect()
@@ -93,6 +110,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        // if (\Gate::denies('delete-question', $question)) {
+        //     abort(403, 'Access denied');    
+        // }
+
+        $this->authorize('delete', $question);
+
         $question->delete();
 
         return redirect()
